Separate menu and reviews to components
